signin/requirements.txt
Flask_Script==2.0.6
Flask_Migrate==2.5.2
flask_restplus==0.13.0
Flask_Bcrypt==0.7.1
Flask==1.1.1
flask_pymongo==2.3.0
PyJWT==1.7.1
